Adjust Charles to new migrate.js files
For now only n:n topology is supported but the fact that it's working is truly marvelous xD
Had to find magical CodeReader.java file and edit it properly for it to start working ;)

emas-lamarckin and emas-baldwin lack some files i.e. common.js
Other configurations work as smoothly as humanly possible.